Prefer redirecting to mirror instead of WG14 URL
The previous commit added some `mirror` keys to documents.yml for
documents that were missing on the WG14 site but have equivalent
versions hosted elsewhere. This commit changes the C18 draft to use the
`mirror` key as well. The app and redirect data generation has been
modified to use the mirror if it is available.

In the future we will be able to select which link to use, for example
by a query string. For now we only use the default. See also issue #25.