Run CI on master branch only
Running CI on push and pull requests to master means that double build
situations will be avoided. For example when doing a PR from the same
repositories' feature branch, both the PR and branch will build.

This prevents CI from running on commits to feature branches but that is
acceptable.