Refactor validation (#34)
Replaces validation using Jest tests with a custom script. The
implementation is a lot less convulated because Jest isn't made for that
kind of thing so the assertions it was doing didn't actually make sense,
even if they worked.

Errors are much more clear now and the output includes line and column.
By comparison, the data validation Jest tests had no location info and
the schema validation would give the path to the error (like `[123].id`)
which for documents.yml is nearly useless since there are so many
entries.

Also changes the YAML library once again because I needed an easy way
to get location info.